requirements.txt
asyncio
python-socketio[asyncio_client]
python-socketio[client]
requests
